+++ src/main/java/com/sdl/selenium/web/WebDriverConfig.java
@@ -1,7 +1,0 @@
-package com.sdl.selenium.web;
-
-/**
- * @deprecated use com.sdl.selenium.utils.config.WebDriverConfig
- */
-public class WebDriverConfig extends com.sdl.selenium.utils.config.WebDriverConfig {
-}