+++ release-notes.md
@@ -1,7 +1,7 @@
 RELEASE NOTES
 -------------
 **Release Notes for Testy 1.8.0-SNAPSHOT**
-- 
+- remove deprecated classes and methods
 
 **Release Notes for Testy 1.7.17**
 - create XPathBuilder